+++ src/main/java/com/acorn/api/controller/contact/ContactCancelController.java
@@ -3,7 +3,7 @@ package com.acorn.api.controller.contact;
 import com.acorn.api.code.common.ApiSuccessCode;
 import com.acorn.api.code.response.ApiResponseBuilder;
 import com.acorn.api.code.response.ApiSuccessResponse;
-import com.acorn.api.dto.contact.requset.ContactCancelReqDTO;
+import com.acorn.api.dto.contact.request.ContactCancelReqDTO;
 import com.acorn.api.service.contact.ContactService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
