+++ src/main/java/com/acorn/api/controller/contact/ContactUpdateController.java
@@ -3,7 +3,7 @@ package com.acorn.api.controller.contact;
 import com.acorn.api.code.common.ApiSuccessCode;
 import com.acorn.api.code.response.ApiResponseBuilder;
 import com.acorn.api.code.response.ApiSuccessResponse;
-import com.acorn.api.dto.contact.requset.ContactUpdateReqDTO;
+import com.acorn.api.dto.contact.request.ContactUpdateReqDTO;
 import com.acorn.api.service.contact.ContactService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
